+++ docs/source/conf.py
@@ -43,10 +43,12 @@ html_theme = 'sphinx_rtd_theme'
 # -- Options for EPUB output
 epub_show_urls = "footnote"
 
+
 def skip(app, what, name, obj, would_skip, options):
     if name == "__init__":
         return False
     return would_skip
+
 
 def setup(app):
     app.connect("autodoc-skip-member", skip)
