+++ docs/source/moduleApi/lsg.rst
@@ -4,8 +4,14 @@ Random walk
 .. automodule:: ffpack.lsg.randomWalk
    :members:
 
-Autoregressive model
+Autoregressive moving average
 --------------------
 
-.. automodule:: ffpack.lsg.autoregressiveModel
+.. automodule:: ffpack.lsg.autoregressiveMovingAverage
+   :members:
+
+Sequence from spectrum
+--------------------
+
+.. automodule:: ffpack.lsg.sequenceFromSpectrum
    :members:
