+++ setup.cfg
@@ -1,3 +1,3 @@
 [flake8]
-ignore = E201,E202,E302,E303,E731,F401,F403,W291,W293
+ignore = E201,E202,E302,E303,E731,F401,F403,W291,W293,W504
 max-line-length = 127
