+++ src/ffpack/lsg/autoregressiveMovingAverage.py
