+++ tests/utils/test_utils_miscUtils.py
@@ -5,7 +5,6 @@ import numpy as np
 import pytest
 
 
-
 ###############################################################################
 # Test centralDiffWeights
 ###############################################################################
@@ -65,7 +64,6 @@ def test_centralDiffWeights_ndivIsTwo_array( Np ):
                                 np.round( expectedWeights, 5 ) )
 
 
-
 ###############################################################################
 # Test derivative
 ###############################################################################
@@ -112,7 +110,6 @@ def test_derivative_nonLinearFunCase2_scalar( x0 ):
     calRst = utils.derivative( f, x0, dx=1e-6 )
     expectedRst = df( x0 )
     np.testing.assert_allclose( calRst, expectedRst )
-
 
 
 ###############################################################################
@@ -239,7 +236,6 @@ def test_gradient_thirdDerivativeCase_funcList( X0 ):
                                               np.round( expectedRst[ i ]( X0 ), 4 ) )
 
 
-
 ###############################################################################
 # Test hessianMatrix
 ###############################################################################
@@ -339,7 +335,6 @@ def test_hessianMatrix_trigonometricCase_funcList( X0 ):
             np.testing.assert_array_almost_equal( 
                 np.round( calRst[ i ][ j ]( X0 ), 4 ), 
                 np.round( expectedRst[ i ][ j ]( X0 ), 4 ) )
-
 
 
 ###############################################################################
